kernel/power: Improve performance of LZO/plain hibernation
Pekka was right - there is no measurable difference in performance,
whether we bind threads to specific CPUs or not. So, I dropped that, it
is not required and makes code simpler.

Other than that, just some code simplifications, variable name cleanups
and buglet removals.

Note that I changed the subject to reflect the fact that
hibernation/thaw without compression also benefits from the patch, from
the improved I/O.

---------------------------------------
Use threads for LZO compression/decompression on hibernate/thaw.
Improve buffering on hibernate/thaw.
v8
In my testing, this improved write/read speed by a factor of 2 to 3.

Signed-off-by: Bojan Smojver <bojan@rexursive.com>